Initial commit
Committing all files from an abandoned side project due to having fulfilled it's purpose - it would be worth someone adding a requirements file at some point in future. The only dependency I can recall is cryptos but there may be a bit of trial and error with a fresh pip environment.